server/pyproject.toml
[project]
name = "real-estate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.1",
    "milvus>=2.3.9",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.3.1",
    "pymilvus>=2.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
]
